+++ src/Router/index.tsx
@@ -12,13 +12,16 @@ import ModifyStudyPage from '@/Pages/Studies/EditStudy';
 import CreateRecruitmentPage from '@/Pages/CreateRecruitment/page';
 import LoginFailPage from '@/Pages/LoginFail';
 import SignUpFailPage from '@/Pages/SignUpFail';
+import MyPage from '@/Pages/MyPage/index';
+import { Saved } from '@/Pages/Saved';
+import { Notifications } from '@/Pages/Notifications';
+
 import Header from '@/Components/Header';
 import Footer from '@/Components/Footer';
 import ErrorBoundary from '@/Components/ErrorBoundary';
 import { EditRecruitmentFetcher } from '@/Pages/EditRecruitment/EditRecruitmentFetcher';
 import { MyPageLayout } from '@/Layout/MyPageLayout';
 import { SettingLayout } from '@/Layout/SettingLayout';
-import { Notifications } from '@/Pages/Notifications';
 
 export const RouterPath = createBrowserRouter([
   {
@@ -57,7 +60,7 @@ export const RouterPath = createBrowserRouter([
           {
             // 회원 정보
             index: true,
-            element: <>TODO</>,
+            element: <MyPage />,
           },
           {
             // 스터디원이 남긴 나의 리뷰
@@ -67,7 +70,7 @@ export const RouterPath = createBrowserRouter([
           {
             // 임시 저장된 글
             path: ROUTES.MYPAGE.SAVED,
-            element: <>TODO</>,
+            element: <Saved />,
           },
           {
             path: ROUTES.MYPAGE.SETTINGS,
